backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
pydantic==2.11.3
pydantic-settings==2.8.1
python-multipart==0.0.20
openai==1.76.0
sqlalchemy==2.0.40
psycopg[binary]==3.2.6
pypdf==5.4.0
python-docx==1.1.2
